requirements.txt
aiohttp==3.8.2
aiosignal==1.3.1
anvil-uplink==0.4.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
bitarray==2.7.4
bytestring-splitter==2.4.1
cached-property==1.5.2
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constant-sorrow==0.1.0a9
constantly==15.1.0
coverage==6.5.0
cryptography==41.0.1
cytoolz==0.12.1
dateparser==1.1.8
decorator==5.1.1
distlib==0.3.6
eip712-structs==1.1.0
eth-abi==4.0.0
eth-account==0.8.0
eth-bloom==2.0.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.9.0b1
eth-typing==3.3.0
eth-utils==2.1.0
exceptiongroup==1.1.1
ferveo @ git+https://github.com/KPrasch/ferveo.git@7a5c5fc8be49894c0affd0da32b80e29e6e1ee8e#subdirectory=ferveo-python
filelock==3.12.0
Flask==2.2.5
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
hendrix==4.0.0
hexbytes==0.3.0
humanize==4.6.0
hyperlink==21.0.0
hypothesis==6.72.1
identify==2.5.22
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.0.3
jsonschema==4.18.0a9
jsonschema-specifications==2023.5.2
lru-dict==1.2.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
maya==0.6.1
mnemonic==0.20
msgpack==1.0.5
msgpack-python==0.5.6
multidict==5.2.0
mypy-extensions==0.4.4
nodeenv==1.7.0
nucypher @ git+https://github.com/nucypher/nucypher@1032a2b5560ef8e5afa02097510f13812943e219
nucypher-core==0.10.0
packaging==23.1
parsimonious==0.9.0
pendulum==3.0.0a1
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.2
prometheus-client==0.16.0
protobuf==4.23.2
py==1.11.0
py-ecc==6.0.0
py-evm==0.7.0a2
py-solc-x==0.10.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pychalk==2.0.1
pycparser==2.21
pycryptodome==3.18.0
pyethash==0.1.27
pyflakes==3.0.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pysha3==1.0.2
pytest==6.2.5
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest-timeout==2.1.0
pytest-twisted==1.14.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
referencing==0.29.0
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.7.1
semantic-version==2.10.0
sentry-sdk==1.23.1
service-identity==21.1.0
six==1.16.0
snaptime==0.2.4
sortedcontainers==2.4.0
tabulate==0.9.0
time-machine==2.9.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
trie==2.1.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
virtualenv==20.22.0
watchdog==2.3.1
web3==6.4.0
websockets==11.0.3
Werkzeug==2.3.4
ws4py==0.5.1
yarl==1.9.2
zipp==3.15.0
zope.interface==6.1a2
